refactor(hew-lsp): extract navigation handlers
