+++ hew-lsp/src/server/handlers/mod.rs
@@ -1,1 +1,2 @@
+pub(super) mod navigation;
 pub(super) mod text_sync;
